+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -150,7 +150,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KQES46MMNA1XF1SPK1JKA867"
+        "01KQEYGM4R9JK5B5TNT2X3NHH1"
     )
     r"""The unique identifier of the data source"""