+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KQES46MPNQBXXN918R179K6Z"
+        "01KQEYGM4ZAAP32ZTHDM2CWDVE"
     )
     r"""The unique identifier of the data source"""